+++ CHANGELOG.md
@@ -16,6 +16,7 @@
 
 ### Chore & Maintenance
 
+- `[jest-changed-files, jest-circus, jest-jasmine2, jest-runner]` Replace `p-limit` with inline concurrency limiter in `jest-util`
 - `[docs]` Update V30 migration guide to notify users on `jest.mock()` work with case-sensitive path ([#15849](https://github.com/jestjs/jest/pull/15849))
 - `[deps]` Update to sinon/fake-timers v15
 - Updated Twitter icon to match the latest brand guidelines.([#15868](https://github.com/jestjs/jest/pull/15869))